uses `read-write-lock` instead of `level-lock`
