+++ index.js
@@ -1,21 +1,21 @@
 var EventEmitter = require('events').EventEmitter
+var ReadWriteLock = require('read-write-lock')
 var uuid = require('random-uuid-v4')
 var ttl = require('tiny-level-ttl')
 var xtend = require('xtend')
 var authenticate = require('./authenticate.js')
 var beginAuthentication = require('./beginAuthentication.js')
 
-var defaultOptions = {
-	tokenGenerator: uuid,
-	tokenTtl: 5 * 60 * 1000, // 5 min
-	tokenTtlCheckIntervalMs: 10 * 1000 // 10 sec
-}
-
 module.exports = function JustLoginCore(tokenDb, options) {
 	if (!tokenDb) {
 		throw new Error("Just Login Core requires a valid levelup database!")
 	}
-	var opts = xtend(defaultOptions, options)
+	var opts = xtend({
+		tokenGenerator: uuid,
+		tokenTtl: 5 * 60 * 1000, // 5 min
+		tokenTtlCheckIntervalMs: 10 * 1000, // 10 sec
+		tokenLock: ReadWriteLock()
+	}, options)
 
 	ttl(tokenDb, {
 		ttl: opts.tokenTtl,
@@ -24,8 +24,8 @@ module.exports = function JustLoginCore(tokenDb, options) {
 
 	var emitter = new EventEmitter()
 
-	emitter.authenticate = authenticate(emitter, tokenDb)
-	emitter.beginAuthentication = beginAuthentication(emitter, tokenDb, opts.tokenGenerator)
+	emitter.authenticate = authenticate(emitter, tokenDb, opts.tokenLock)
+	emitter.beginAuthentication = beginAuthentication(emitter, tokenDb, opts.tokenLock, opts.tokenGenerator)
 
 	return emitter
 }
